+++ src/updater/index.ts
@@ -2,7 +2,6 @@ export {
   steamcmdDownloadGame,
   steamcmdDownloadSelf,
   steamcmdInteractiveLogin,
+  steamcmdCheckUpdate,
   hasSteamUpdate,
-  getLocalSteamDepot,
-  getRemoteSteamDepot,
 } from './steam';